requirements.txt
chainlit>=1.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
tenacity>=8.2.0
